Fix: 로그인/비로그인시 아이콘으로 상태 표시
